interview_project_PYTHON/requirements.txt
appium-python-client==4.0.0
attrs==25.3.0
certifi==2025.8.3
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pip==23.0.1
pysocks==1.7.1
selenium==4.35.0
setuptools==58.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
